CloudFunction_On_Local/requirements.txt
Flask==2.1.2
google-cloud-pubsub==2.7.0
faker==8.8.2
functions-framework==3.0.0
